# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Hangman API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScoreEntryMode =
  (typeof ScoreEntryMode)[keyof typeof ScoreEntryMode];

export const ScoreEntryMode = {
  solo: "solo",
  daily: "daily",
} as const;

export interface ScoreEntry {
  id: number;
  userId: string;
  userName: string;
  userAvatar?: string | null;
  score: number;
  streak: number;
  word: string;
  mode: ScoreEntryMode;
  won: boolean;
  createdAt: string;
}

export type SubmitScoreBodyMode =
  (typeof SubmitScoreBodyMode)[keyof typeof SubmitScoreBodyMode];

export const SubmitScoreBodyMode = {
  solo: "solo",
  daily: "daily",
} as const;

export interface SubmitScoreBody {
  score: number;
  streak: number;
  word: string;
  mode: SubmitScoreBodyMode;
  won: boolean;
  userName: string;
  userAvatar?: string | null;
}

export interface LeaderboardStats {
  totalGames: number;
  topScore: number;
  averageScore: number;
  totalPlayers: number;
}

export interface WordList {
  words: string[];
  category?: string;
}

export interface DailyChallenge {
  word: string;
  date: string;
  hint?: string | null;
}

export type GetLeaderboardParams = {
  limit?: number;
  mode?: GetLeaderboardMode;
};

export type GetLeaderboardMode =
  (typeof GetLeaderboardMode)[keyof typeof GetLeaderboardMode];

export const GetLeaderboardMode = {
  solo: "solo",
  daily: "daily",
  all: "all",
} as const;

export type GetWordsParams = {
  category?: string;
};