# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Hangman API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns the top scores for the leaderboard
 * @summary Get leaderboard
 */
export const getLeaderboardQueryLimitDefault = 20;
export const getLeaderboardQueryModeDefault = `all`;

export const GetLeaderboardQueryParams = zod.object({
  limit: zod.coerce.number().default(getLeaderboardQueryLimitDefault),
  mode: zod
    .enum(["solo", "daily", "all"])
    .default(getLeaderboardQueryModeDefault),
});

export const GetLeaderboardResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  userName: zod.string(),
  userAvatar: zod.string().nullish(),
  score: zod.number(),
  streak: zod.number(),
  word: zod.string(),
  mode: zod.enum(["solo", "daily"]),
  won: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetLeaderboardResponse = zod.array(GetLeaderboardResponseItem);

/**
 * @summary Submit a score
 */
export const SubmitScoreBody = zod.object({
  score: zod.number(),
  streak: zod.number(),
  word: zod.string(),
  mode: zod.enum(["solo", "daily"]),
  won: zod.boolean(),
  userName: zod.string(),
  userAvatar: zod.string().nullish(),
});

/**
 * @summary Get current user's scores
 */
export const GetMyScoresResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  userName: zod.string(),
  userAvatar: zod.string().nullish(),
  score: zod.number(),
  streak: zod.number(),
  word: zod.string(),
  mode: zod.enum(["solo", "daily"]),
  won: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetMyScoresResponse = zod.array(GetMyScoresResponseItem);

/**
 * Returns aggregate stats — total games played, top scorer, average score
 * @summary Get leaderboard statistics
 */
export const GetLeaderboardStatsResponse = zod.object({
  totalGames: zod.number(),
  topScore: zod.number(),
  averageScore: zod.number(),
  totalPlayers: zod.number(),
});

/**
 * Returns a set of words for the game
 * @summary Get word list for game
 */
export const GetWordsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
});

export const GetWordsResponse = zod.object({
  words: zod.array(zod.string()),
  category: zod.string().optional(),
});

/**
 * Returns the current daily challenge word (same for all users)
 * @summary Get today's daily challenge word
 */
export const GetDailyChallengeResponse = zod.object({
  word: zod.string(),
  date: zod.string(),
  hint: zod.string().nullish(),
});

/**
 * @summary Get today's daily challenge leaderboard
 */
export const GetDailyLeaderboardResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  userName: zod.string(),
  userAvatar: zod.string().nullish(),
  score: zod.number(),
  streak: zod.number(),
  word: zod.string(),
  mode: zod.enum(["solo", "daily"]),
  won: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetDailyLeaderboardResponse = zod.array(
  GetDailyLeaderboardResponseItem,
);