# lib/api-zod/src/generated/types/dailyChallenge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Hangman API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyChallenge {
  word: string;
  date: string;
  hint?: string | null;
}