# lib/api-zod/src/generated/types/leaderboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Hangman API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardStats {
  totalGames: number;
  topScore: number;
  averageScore: number;
  totalPlayers: number;
}