# lib/api-zod/src/generated/types/scoreEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Hangman API
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreEntryMode } from "./scoreEntryMode";

export interface ScoreEntry {
  id: number;
  userId: string;
  userName: string;
  userAvatar?: string | null;
  score: number;
  streak: number;
  word: string;
  mode: ScoreEntryMode;
  won: boolean;
  createdAt: Date;
}