# lib/api-zod/src/generated/types/submitScoreBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Hangman API
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitScoreBodyMode } from "./submitScoreBodyMode";

export interface SubmitScoreBody {
  score: number;
  streak: number;
  word: string;
  mode: SubmitScoreBodyMode;
  won: boolean;
  userName: string;
  userAvatar?: string | null;
}